Initial commit: a simple routing engine
For this commit, I have routing working between tunnel interfaces in
seperate network namespaces on Linux.  There is tooling in the
repository to aid in constructing these tunnel interfaces and when
started (as hardcoded) the main program creates a routed link between
the default network namespace and the `blue` namespace.

This has been tested with ping, curl, and iperf and seems to be
relatively stable locally, achieving ~15Mb/s with all print statements
and ~400Mb/s with none.

Next thing is probably to think about a configuration outside of just
hardcoding routing rules, and I also want to move on to things like NAT
eventually I think.