routing: add local routing and icmp reply
Added the facility for a packet to be marked as local to the router.
This allows for things like ICMP responces or status queries to happen.
These packets still pass through the normal routing engine, and indeed
replies are dispatched by showing packets into the routing pipeline.
This is super elegant if you ask me, and is right in line with what I
was hoping to be able to do.

ICMP replies to the local IPs defined in the routes table (the (local,)
rules) do work, but something is wrong with the checksum as the pinging
program complains about a bad checksum.