better ideas
The key insight tonight was that routes don't need arbitrarily
complicated destinations or a "local packet" notion.  All routes can
poinr to a struct representing an interface, and packets destined for
localhost will go into the struct representing the local interface,
regardless of their source.

The "send a packet" function pointer on a real interface needs to engage
the driver to send bytes on the wire, and the "send a packet" function
pointer on the local interface needs to look up a socket, update state
there, put data in a buffer, and wake up a thread.